Move the Record::Asset specs into a shared example
This way the Asset specs can be shared with Layouts
